refactor(glplayer): let the builder to be a standalone class
